installer: add alternative check for real rootfs partition
The device symlink that systemd installs for systemd.volatile=overlay
setups isn't that helpful when using LUKS for the rootfs partition,
as it points to the underlying physical device rather than the
devmapper device we need.

So for LUKS rootfs setups on tegra platforms, the generator script
that sets up the cryptfs and mount units for the rootfs installs
a symlink in another location to make it easier to find the underlying
rootfs device for the overlay rootfs setup.  Use that if it's present.

Signed-off-by: Matt Madison <matt@madison.systems>